fix: migrate postgres tests to their own file
